Update resolve functions to pass down the error code
If cil_symtab_get_node fails because the item didn't exist (e.g. not a
memory error or something), then we need to pass that error all the way
down, because there is a now a difference between something failing
(SEPOL_ERR) and we not being able to resolve a something (SEPOL_ENOENT).