Don't bother disabling an optional if it's parent optional is disabled
An parent optional wont always get disabled by the time we are in a
child optional, for example:

(optional foo
	(optional bar
		(type baz))
	(allow a b file (read))

foo will get disabled when the allow statement fails to resolve. By that
time, we are already out of (optional bar) and won't be able to disable
it. So we'll do that elsewhere.