Track going into macros/tunableifs correctly when building the ast
Instead of looking at the ast flavor, we need to be looking at the parse
tree string for determining if the parse tree parser is going inside a
block. This error caused the parser to think it was entering a tunableif
block multiple times, so it thought it was in a tunable if when it
wasn't.

We should still use the ast flavor for determining when we're going out
of a block.