Actually fix tracking going into macros/tunables
Commit 02b2a1154c78075e797a830f7bf115ce4e336489 didn't really fix the
problem. We can't use first child helper at all for tracking because
that is the first child in the parser tree. We really need to look at
the first child of the ast tree, which we can only get access to in the
main node function. So do the checking there.