Do not count duplicate types and roles
Duplicate types and roles are allowed, like when inherited from blocks
and macro calls. Although multiple AST nodes for the same type/role
exist when this happens (for diagnostic purposes), the AST nodes share
the same datum. So when calculating a type or role value in cil_post, we
need to do so only once per datum. The standard CIL way to do this is to
check that the current node is the first one in the datums list of
nodes.

Signed-off-by: Steve Lawrence <slawrence@tresys.com>
Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>