Do not print duplicate levels in file_contexts
If the high and low level are identical, then only one level will
be exported to file_contexts.

For example:

/usr/bin/ddd    policy.system_u:policy.system_r:kernel_t:s0.c0-s0:c0

becomes:

/usr/bin/ddd    policy.system_u:policy.system_r:kernel_t:s0

Additionally, we have moved the sensitivities and categories out of
the block in the test cil policy. Sensitivities and categories are not
allowed inside blocks because it causes issues when parsing
file_contexts.

Signed-off-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>