Do not insert unused typeattributes into the binary
Typeattributes, especially those used in neverallow rules that generally
include a lot of types, can cause performance issues, even when not used
by any statements. So if we detect that a typeattribute is not used or
is only used by statements that are not added to the binary, such as
neverallow rules, then do not add them into the binary.

Signed-off-by: Steve Lawrence <slawrence@tresys.com>