Move symtab from cil_db to root ast node
Moving the symtab to the root ast node removes the requirement for using
the cil_db to get the symtab.

Since cil_db is no longer needed for many of the cil_build functions, we
can remove the argument. Because we might need the cil_db in the future,
we are making it unused rather than outright deleting.

Signed-off-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>