Cleanup unused symtab_datum from disabled optionals
Children of optional AST nodes are no longer disabled by using a
'disabled' state. The children of a disabled optional are instead
completely removed from the AST. If the symtab_datum the node was pointing to
has no more AST nodes in its node list, then the datum is destroyed.

Removed state flag from symtab_datums. Optionals will now keep track if
they are disabled or enabled.

Add an fqn member to cil_symtab_datum. This allows for symtab lookups to
work properly when destroying symtabs because the fqn is not the correct key
to use for the symtab lookups.

Move logic to remove an item from a cil_list to a helper function for
clarity.

Signed-off-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>