Do not associate the object_r role with users
The object_r role is implicitly associated with all users. However, CIL
requires that roles, even object_r, must be associated with users so
that we can validate things like contexts, where users and roles are
used together. So when generating the binary policy, we set a bit for
the user/object_r role association. However, because this association is
implied, this actually caused some userspace tools to change behavior,
and in some cases break (e.g. semanage user -l). So to maintian tooling
consistency, do not explicity associate any users with object_r, and
instead rely on the implicit association existing.

Signed-off-by: Steve Lawrence <slawrence@tresys.com>