Modify how blocks are copied into blockinherits
This patch allows consistent behavior for blockinherits with
identical names. Two passes will now occur before blockinherits
have their contents copied correctly.

In the first pass, all blockinherit nodes will be iterated. Each
blockinherit will be linked to the block that will need to be
copied. Each block will contain a list of blockinherit nodes
that inherit the block.

In the second pass, all block nodes are iterated. If a block is
inherited by any blockinherits, and if a block is the first in the
block datums node list i.e. the block is the original block, and
not a block merged in from another blockinherit, then the children
of the block will be copied as children of the blockinherit. When a
blockinherit node is copied, a new node is added to any list inside the
block that inherited it. This will prevent any ordering
issues from occuring.

Signed-off-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>