Reimplemented the code to fully qualify names.
The qualified name will be based on which block's symtab the datum
is in rather than on the location of the node where the datum is
declared.

This changes how declarations in macros have their names qualified.
Before their names were qualified based on where the macro was called,
but now they are qualified based on where the macro is defined.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>