Store file paths in the string pool
Anything we are given from the user should be copied to avoid potential
issues where a user changes something after we store it. So when we
receive a file path from the user in cil_paser, add the path to the
string pool and use that, forgetting about the string the user gave us.

Signed-off-by: Steve Lawrence <slawrence@tresys.com>
Reviewed-by: Yuli Khodorkovskiy <ykhodorkovskiy@tresys.com>