Changed how datums are handled when destroying a tree.
This change adds a pointer in the datum to its symtab.

When a tree node containing a datum is destroyed, its node will be
removed from the datum's node list. If the node list is now
empty, then the node list is destroyed and the datum is removed
from the symtab (if the symtab has not been destroyed already).

When destroying a symtab, each datum in the symtab will have its
symtab pointer set to NULL.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>